Replace the own cars list in the main ui with buttons
Makes it more obvious that they are clickable items. To get empty
buttons with borders custom drawables have to be used. To get the
colored button (unpark button) to look like the non colored (park
button) that version is also done to get similar buttons.

Unfortunately the ripple effect gets canceled when the drawable
get's changed during the effect. But that's something we have
to live with.