Make sure all resources are always initialized
It's not possible to only init things in the main activity. If the app have
been killed in the background and restored to another activity later it
will crash due to missing initializations.

Create a ParkApp singleton that initializes and holds all critical resources
and can be used to get the app context from anywhere.