Don't use Google Play services if it doesn't exist
Do a simple best effort approach and avoid using features that obviously don't
work without google play services such as signing in and dynamic links. Other
features like Firebase and Crashlytics might or might not work fully but the app
is at least usable and not having google play services should be very rare.

For the same reason the error message shown when google play services is
not present is a message provided by Google  for the sake of simplicity. It's not
completely correct, but good enough for something that more or less no one
should ever see.