Fix bug with wrong placeholder after closing app with back key
Keeping state in the NetworkManager (which is global) which is meant for the
ParkActivity is not a good idea. When the ParkActivity is destroyed and recreated
it's internal state is reset, the global state of the Network manager is not reset.

So move the state from the NetworkManager to the ParkActivity instead so the
state is kept consistently.