Don't use plurals for park status
Plurals is based on the phone language and not the language used in the UI so if the
phone language is set to Japanese the "other" rule will be used even when there is
only one item. So plurals might get wrong when using unlocalized languages that
fallback to English.

This case is not even a case where plurals should be used in the first place so stop
using it.