Temporarily(?) disable the wait list feature
Removes the possibility to log in and grays out the park button
when the garage is full instead of showing the wait button.

There is no backend any more, hopefully a new backend will be
introduced again at some point and then this can be restored.

But with everyone working from home and the garage being empty
all the time there is little value of a wait list feature
right now.