Upgrade gradle to 7.0
Required a bump of robolectric to 4.4 to work at all. Then a bunch
of changes were required to make test work with the new paused looper
http://robolectric.org/blog/2019/06/04/paused-looper/

Their documentation sucks as allways so very hard to figure out what
needs to be done, this is just by trial and error without much understanding.

Had also to disable the dynamic link / play services code when running
under test to get rid of tests that fails when running all tests